+++ src/main/php/text/ical/Calendar.class.php
@@ -9,6 +9,9 @@ class Calendar implements Object {
 
   private $method, $prodid, $version, $event, $timezone;
 
+  /** @return text.ical.Event[] */
+  public function events() { return (array)$this->event; }
+
   /**
    * Write this object
    *